Move sidebar menu buttons into separate file
